+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## 2021-07-13 (v1.0.14)
+### Added
+- 'animationOff' - turn off the progress animation
 
 ## 2021-07-13 (v1.0.13)
 ### Added